tolkar-factory/requirements.txt
# TOLKAR Zero@Factory - Python Dependencies
# Production Intelligence Platform for Smartex Machines
# Phase 9.5 - Production Ready

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
pyjwt==2.8.0
bcrypt==4.0.1
cryptography==41.0.7
passlib==1.7.4

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# PostgreSQL Support (optional - for production)
psycopg2-binary==2.9.9
asyncpg==0.29.0

# MySQL Support (optional - alternative)
mysql-connector-python==8.2.0
aiomysql==0.2.0

# Utilities
python-multipart==0.0.6
email-validator==2.1.0

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Production Dependencies
python-json-logger==2.0.7
sentry-sdk==1.38.0

# Optional: Redis for caching (future enhancement)
redis==5.0.1
aioredis==2.0.1

# Optional: Celery for background tasks (future enhancement)
celery==5.3.4

# Optional: WebSocket support (future enhancement)
websockets==12.0

# Optional: File upload support
aiofiles==23.2.1

# Optional: Excel/CSV export support
openpyxl==3.1.2
pandas==2.1.3

# Optional: PDF generation support
reportlab==4.0.7

# Optional: Image processing
pillow==10.1.0

# Optional: QR code generation
qrcode==7.4.2

# Optional: API documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Optional: Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Optional: Performance monitoring
py-spy==0.3.14
memory-profiler==0.61.0